Added cleanup of files not listed in filest
Now, if a filelist is present it will not only be used to calculate the
hashsum of the files, but also to remove files that don't belong.

But wasn't the point of the filelist *not* to remove added files?
- No, it was to make sure that patching wouldn't fail due to added
files.
This approach is reasonable because the same effect (deleted files)
would happen if the patch failed and a full update was downloaded
instead.

I have also included some tests for the filelist cases of patching.